\name{EpivizBlockDataPack-class}
\docType{class}

\alias{class:EpivizBlockDataPack}
\alias{EpivizBlockDataPack-class}

\title{'EpivizBlockDataPack' objects}

\description{Package of block data for transfer to web app.}

\section{Fields}{
  \describe{
    \item{\code{data}:}{(list) data is kept here for packaging. Slot names correspond to measurement ids.}
  }
}

\author{Hector Corrada Bravo}

