\name{EpivizBpDataPack-class}

\alias{class:EpivizBpDataPack}
\alias{EpivizBpDataPack-class}

\title{'EpivizBpDataPack' objects}

\description{Package of base-pair level data for transfer to web app.}

\section{Fields}{
  \describe{
    \item{\code{min}:}{(numeric) minimum of data range.}
    \item{\code{max}:}{(numeric) maximum of data range.}
    \item{\code{data}:}{(list) data entry for each measurement.}
  }
}

\author{Hector Corrada Bravo}


