% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizChartMgr-class.R
\docType{class}
\name{EpivizChartMgr-class}
\alias{EpivizChartMgr}
\alias{EpivizChartMgr-class}
\title{Class providing chart manager for epiviz app}
\description{
Class providing chart manager for epiviz app
}
\section{Methods}{

\describe{
\item{\code{is_server_closed()}}{Check if underlying server is closed, <logical>}
}}

