\name{EpivizDataCache-class}
\docType{class}

\alias{class:EpivizDataCache}
\alias{EpivizDataCache-class}

\title{
Virtual class for data cache used by Epiviz.
}
\description{
Manage data cache from external resources (e.g., BigWig files) for use by epiviz queries. 
}
\details{
Manage data cache from external resources (e.g., BigWig files) for use by epiviz queries. 
}
\author{
Hector Corrada Bravo <hcorrada@gmail.com>
}
