\name{EpivizTrackData-class}
\docType{class}

\alias{class:EpivizTrackData}
\alias{EpivizTrackData-class}

\title{'EpivizTrackData' objects}

\description{Data container for data displayed in tracks}

\details{This class extends \code{EpivizData-class} directly.
The data containing \code{object} is constrained to be of class 
\code{\link[GenomicRanges]{GIntervalTree}}.}

\author{Hector Corrada Bravo}

