\name{EpivizWigCache-class}
\docType{class}

\alias{class:EpivizWigCache}
\alias{EpivizWigCache-class}

%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Cache for BigWig files
}
\description{
Manages a data cache for external BigWig files.
}
\details{
Manages a data cache for external BigWig files.
}
\author{
Hector Corrada Bravo <hcorrada@gmail.com>
}
