\name{EpivizWigData-class}

\alias{class:EpivizWigData}
\alias{EpivizWigData-class}

%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Container for data read from external BigWig files
}
\description{
Container for data read from external BigWig files
}
\details{
Container for data read from external BigWig files
}
\author{
Hector Corrada Bravo <hcorrada@gmail.com>
}
