% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restartEpiviz.R
\name{restartEpiviz}
\alias{restartEpiviz}
\title{Restart epiviz app connection and workspace.}
\usage{
restartEpiviz(file)
}
\arguments{
\item{file}{(character) The name of the file that holds the EpivizApp object to be restarted, ending in .rda.}
}
\value{
An object of class \code{\link{EpivizApp}}
}
\description{
Restart epiviz app connection and workspace.
}
\examples{
# see package vignette for example usage
app <- restartEpiviz(file="app.rda")

}

