% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizChart-class.R
\docType{methods}
\name{show,EpivizChart-method}
\alias{show,EpivizChart-method}
\title{Print information about EpivizChart}
\usage{
\S4method{show}{EpivizChart}(object)
}
\arguments{
\item{object}{an \code{\link{EpivizChart}} object}
}
\description{
Print information about EpivizChart
}

