% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startEpiviz.R
\name{startEpiviz}
\alias{startEpiviz}
\title{Start epiviz app and create \code{\class{EpivizApp}} object to manage connection}
\usage{
startEpiviz(port = 7312L, localURL = NULL, useDevel = FALSE,
  standalone = FALSE, staticSitePath = "", chr = "chr11",
  start = 99800000, end = 103383180, debug = FALSE, workspace = NULL,
  scripts = NULL, gists = NULL, openBrowser = TRUE,
  daemonized = .epivizrCanDaemonize(), verbose = FALSE,
  nonInteractive = FALSE, tryPorts = FALSE)
}
\arguments{
\item{port}{<integer> The port to open server to}

\item{localURL}{<character> URL to open}

\item{useDevel}{<logical> Use the epiviz-dev app}

\item{standalone}{<logical> Is it running standalone through R}

\item{staticSitePath}{<character> Directory to serve standalone application}

\item{chr}{<character> Chromosome to load at epiviz app start}

\item{start}{<integer> Starting location to load at epiviz app start}

\item{end}{<integer> Ending location to load at epiviz app start}

\item{debug}{<logical> Use debug user for epiviz app workspace storage}

\item{workspace}{<character> Id of workspace to load on epiviz app start}

\item{scripts}{<character> Path to JS scripts to include on epiviz app start}

\item{gists}{<character> Vector of github gist ids with JS code to include on epiviz app start}

\item{openBrowser}{<logical> Open browser at URL through this function call}

\item{daemonized}{<logical> Run app in non-blocking mode}

\item{verbose}{<logical> Print verbose information}

\item{nonInteractive}{<logical> Only used internally}

\item{tryPorts}{<logical> Try multiple ports until finding an open port when opening WebSocket connection}
}
\value{
An object of class \code{\link{EpivizApp}}
}
\description{
Start epiviz app and create \code{\class{EpivizApp}} object to manage connection
}

