% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizChart-class.R
\docType{class}
\name{EpivizChart-class}
\alias{EpivizChart}
\alias{EpivizChart-class}
\title{Class encapsulating a chart in epiviz app}
\description{
Class encapsulating a chart in epiviz app
}
\section{Methods}{

\describe{
\item{\code{get_app_id(id)}}{Get chart's id assigned by epiviz app.}

\item{\code{get_id()}}{Get chart's id assigned by chart manager}

\item{\code{is_connected()}}{Returns \code{TRUE} if chart is connected to a chart on epiviz app.}

\item{\code{set_app_id(id)}}{Set chart's id assigned by epiviz app.}

\item{\code{set_id(id)}}{Set chart's id assigned by chart manager.}
}}

