% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startEpiviz.R
\name{startEpiviz}
\alias{startEpiviz}
\title{Start epiviz app and create \code{\link{EpivizApp}} object to manage connection}
\usage{
startEpiviz(use_devel = FALSE, chr = "chr11", start = 99800000,
  end = 103383180, debug = FALSE, workspace = NULL, scripts = NULL,
  gists = NULL, open_browser = TRUE, non_interactive = FALSE, ...)
}
\arguments{
\item{use_devel}{(logical) use the devel epiviz application server (http://epiviz-dev.cbcb.umd.edu).}

\item{chr}{(character) chromosome to browse to on app startup.}

\item{start}{(integer) start location to browse to on app startup.}

\item{end}{(integer) end location to browse to on app startup.}

\item{debug}{(logical) start epiviz app in debug mode.}

\item{workspace}{(character) a workspace id to load in the epiviz app on startup.}

\item{scripts}{(character) URLs for JavaScript plugin scripts to be imported when epiviz is loaded (see \url{http://epiviz.cbcb.umd.edu/help} for details).}

\item{gists}{(character) Ids for github gists (\url{http://gist.github.com}) containing JavaScript plugin scripts to
be imported when epiviz is loaded (see \url{http://epiviz.cbcb.umd.edu/help} for details).}

\item{open_browser}{(logical) browse to the epiviz URL before exiting function.}

\item{non_interactive}{<logical> run in non-interactive mode, for development purposes only.}

\item{...}{additional parameters passed to \code{\link[epivizrServer]{createServer}}.}
}
\value{
An object of class \code{\link{EpivizApp}}
}
\description{
Start epiviz app and create \code{\link{EpivizApp}} object to manage connection
}

