% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette_data.R
\docType{data}
\name{tcga_colon_expression}
\alias{tcga_colon_expression}
\title{Example exon-level RNAseq data from TCGA project for epivizr vignette.}
\format{
A \code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::RangedSummarizedExperiment}} object with 12,800 rows (exons) and 40 samples.
\describe{
\item{\code{assay(tcga_colon_expression)}}{exon-level count matrix}
\item{\code{colData(tcga_colon_expression)}}{a \code{DataFrame} containing sample information.
Normal/Tumor status is given in column \code{sample_type}}
}
}
\source{
TCGA project: \url{https://tcga-data.nci.nih.gov/tcga/}
}
\usage{
data(tcga_colon_expression)
}
\value{
A \code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::RangedSummarizedExperiment}} object.
}
\description{
A \code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::RangedSummarizedExperiment}} object containing exon-level
counts from RNAseq data for colon tumor and normal tissue from the TCGA project.
Only exons in human chromosome 11 are included.
}
